Style changes: Moves, renames, and reorders.
Like previous PR.
