+++ sphinx_imgur/utils.py
@@ -1,0 +1,19 @@
+"""Helpers."""
+from sphinx.errors import SphinxError
+
+
+class ImgurError(SphinxError):
+    """Non-configuration error. Raised when directive has bad options."""
+
+    category = "Imgur option error"
+
+
+def is_true(argument):
+    """Return True if argument is the word "true" case insensitive.
+
+    :param str argument: Argument/value passed to directive's option.
+
+    :return: Is argument true/True/TRUE.
+    :rtype: bool
+    """
+    return argument.lower() == "true"
